backend/requirements.txt
# FastAPI Core Dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.4.2
pydantic-settings==2.0.3
python-multipart==0.0.6
python-dotenv==1.0.0

# Database Dependencies
sqlalchemy==2.0.23
asyncpg==0.29.0
psycopg2-binary==2.9.9
alembic==1.12.1

# Authentication & Security
PyJWT==2.8.0
cryptography==41.0.7
passlib[bcrypt]==1.7.4

# Django Dependencies (for secondary service)
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
djangorestframework-simplejwt==5.3.0

# HTTP Client & Utilities
httpx==0.25.2
requests==2.31.0
aiofiles==23.2.1

# Caching & Background Tasks
redis==5.0.1

# Email Support
aiosmtplib==3.0.1

# API Documentation Enhancement
fastapi-limiter==0.1.6

# Development Dependencies
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
pytest-mock==3.12.0
black==23.10.1
isort==5.12.0
flake8==6.1.0
mypy==1.7.0

# Database Tools
sqlacodegen==3.0.0

# Monitoring & Logging
structlog==23.2.0

# Security Enhancements
python-decouple==3.8

# File Processing
openpyxl==3.1.2
python-dateutil==2.8.2

# Testing Utilities  
factory-boy==3.3.0
faker==20.1.0